Fixes per discussion with @Ocramius * Change EventListener class to EventLoggingListenerAggregate * Removed version from composer.json * Pass the Event object to collectors from the listener * Create interface and provider class for event context to remove that logic from listener; Class used in collectors
I'm not storing the Event object in the existing collectors' $this->data array as that would be overkill for what they are doing. If additional data is desired to be displayed, it can be pulled directly from the Event object in collectEvent() or functions added to EventContextProvider and then stored in the $this->data array. By giving the collectors access to the Event object, third-party collectors can do as they wish based on their own requirements.